fix: argument parsing consistency
The Go `flag` package that's used to parse command flags will stop at
the first non-flag argument. As the broker name is a mandatory non-flag
argument, it's complex to allow flag arguments to be specified either
before or after the broker name. So we have decided that the broker
name should always come first, followed by optional flags. The
implementation, tests, help text and README have been updated to
consistently follow this pattern.

[#182426367](https://www.pivotaltracker.com/story/show/182426367)