feat: improve textual output
[#185716266](https://www.pivotaltracker.com/story/show/185716266)

This commit doesn't add any of the required improvements.
It focuses on modifying function signatures to accommodate
the required improvements more cleanly in subsequent commits.

This commit alone is able to build but fails to run the tests
because autogenerated fakes got out of sync. I'll regenerate
these fakes in the following commit to simplify revision.