fix: log error details when an error happens upgrading an instance (#98)
Error messages returned by CAPI were ignored and just the http error
status was logged. When the error happend during the actual upgrade
execution in the broker this didn't matter as the error would appear in
the last operation for that service instance. However when there is an
error prior to reaching the broker (e.g. Plan Inaccessible), or prior to
an actual upgrade inside the broker, last_operation for the service
instance remains unchanged and hence the error message was lost, making
troubleshooting impossible through the plugin, and having to restort to
the `cf upgrade-service` command to understand the issue.

[#186169043](https://www.pivotaltracker.com/story/show/186169043)