chore: remove the loop inside of another loop
The loop in the function was executed for every service instance,
so up to 5000 times. Probably, the number of plans is small, and we
believe it is more efficient a map lookupd.
Because we populate the instance with the plan, we no longer needed some fields.
We simplify the query params and reduce network band usage.

[#186597655](https://www.pivotaltracker.com/story/show/186597655)