feat: exit with error if any instance failed update
Previously the plugin would log a summary with the counts of
upgraded, skipped and failed instances plus the details for any
failed instances for users to debug and take action. However, if there
were instances that failed the upgrade, the exit code would still be 0
making it difficult to find out that there were problems upgrading.

This change is to return a non-zero exit code after printing the
summary, when there are one or more instances that failed upgrade.
This makes it easier to detect issues upgrading.

[#184640225](https://www.pivotaltracker.com/story/show/184640225)