feat: remove incorrect method to improve readability
The method `performCheckUpToDate` is intended to execute an
action when there are instances not up-to-date. It turned out
that the method was a dummy method with minimal functionality.
The name of the method was causing problems in understanding
the logic. We prefer to remove it now and perform a refactor
later to improve the modularity of the code.

[#186528305](https://www.pivotaltracker.com/story/show/186528305)