chore: factor out -check-deactivated-plans
The upgrader.Upgrade() function has closely coupled code relating to
multiple distinct actions. As such it's complex and hard to maintain. We
should resolve this by applying the single responsibility principle. The
upgraded.Upgrade() function should have the single responsibility of
being a decider that can call other functions that have single
responsibilies.

Note, this assumes that
https://github.com/cloudfoundry/upgrade-all-services-cli-plugin/pull/130
has been merged, and we might not want to do that.

Note that the new checkDeactivatedPlans() function duplicates some logic
from Upgrade(). This is deliberate as it will allow the logic to be
changed subsequently. For example as implemented the code will list
service instances even when there are no deactivated plans, and
separating the code paths allows this to be optimised away in a future
refactor.