test: integration test for minimum version detection (#211)
The testing in this plugin is mainly of individual components
and not of the system as a whole. This commit introduces some
testing of the plugin as a whole, by creating a fake CAPI backend
and running the plugin via the CF CLI. This testing is still fast
and reliable, and tests behavior rather than implementation: allowing
for refactoring of the internals with low risk of changing behavior.