feat: -ignore-instance-errors flag
The actions -check-deactivated-plans, -check-up-to-date and -min-version-required
will exit with a non-zero code if they find matching service instances. When the
-json flag was added, behavior was changed so that the exit code would always be
zero when using JSON output. The idea was that you might want to pipe the JSON
output to `jq` or a similar processor, and having a failure might not be helpful.
But it was pointed out that you might want to do this with text output too, and
having a change of exit code implicit in the -json flag could be unexpected.
Hence the new -ignore-instance-errors which allows you to control whether finding
instances will result in a non-zero exit code.