feat: ability to set polling interval (#220)
* feat: -ignore-instance-errors flag (#221)

The actions -check-deactivated-plans, -check-up-to-date and -min-version-required
will exit with a non-zero code if they find matching service instances. When the
-json flag was added, behavior was changed so that the exit code would always be
zero when using JSON output. The idea was that you might want to pipe the JSON
output to `jq` or a similar processor, and having a failure might not be helpful.
But it was pointed out that you might want to do this with text output too, and
having a change of exit code implicit in the -json flag could be unexpected.
Hence the new -ignore-instance-errors which allows you to control whether finding
instances will result in a non-zero exit code.

* act on review comments:

- made it clearer that --instance-poll-interval is used as a test helper
  in integration tests, rather then being tested
- added a unit test for the polling interval
- renamed fakeCCAPI to ccapiClient because it's not a test fake - it's
  the system under test

* fix linting and reduce test flakes